Agtor API

Agtor.Climate — Type

Serves as an interface to climate data

Agtor.EconManager — Type

An 'economically rational' crop farm manager.

Water is applied for optimal farm profitability based on soil water deficit, crop water requirements, and cost of water application.

Agtor.Infrastructure — Type

Represents generic farm infrastructure.

Agtor.Irrigation — Type

On-farm irrigation infrastructure component

Agtor.Pump — Type

Pump(pump_efficiency=0.7, cost_per_kW=0.28, derating=0.75)

An on-farm pump to access water.

Agtor.RigidManager — Type

A rigid farm manager.

Water is applied by a pre-determined quantity rather than optimal profitability. Uses surface water first, then groundwater.

Agtor.add_prefix! — Method

Modify AgParameter name in place by adding a prefix.

Agtor.associate_managers — Method

associate_managers(managers, zones)

Map a given list of managers to a list of zones.

Example

OptimizingManager = EconManager("optimizing")
manage_zones = associate_managers([OptimizingManager], [[:Zone_1, :Zone_2]])

Arguments

managers : list-like, of managers

zones : list-like of a list of zone names to associate a manager with

Returns

Tuple of tuples of managers and the names of zones they manage

Agtor.collate_results! — Method

collate_results!(fn_pattern::String, main_fn::String)::Nothing

Collate individually saved .jld2 files from distributed runs into a single JLD2 file.

Agtor.collate_results — Method

collate_results(fn::String, group::String, target::Array{String})::DataFrame
collate_results(data::Dict, group::String, target::Array{String})::DataFrame
collate_results(fn::String, group::String, target::String)::DataFrame
collate_results(data::Dict, group::String, target::String)::DataFrame

Collate a results from across scenario runs into an individual DataFrame.

Agtor.collect_agparams! — Method

collect_agparams!(dataset::Dict, mainset::Dict)::Nothing

Extract parameter values from Dict specification and store in a common Dict.

Agtor.collect_agparams! — Method

Extract parameter values from AgComponent.

Agtor.collect_agparams! — Method

collect_agparams!(dataset::Union{Array, Tuple}, store::Array; ignore::Union{Array, Nothing}=nothing)::Nothing
collect_agparams!(dataset::Dict, store::Array; ignore::Union{Array, Nothing}=nothing)::DataFrame

Agtor.collect_agparams — Method

collect_agparams(dataset::Dict; ignore::Union{Array, Nothing}=nothing)::DataFrame

Extract parameter values from Dict specification.

Agtor.collect_results — Method

Collect model run results for a Basin

Agtor.collect_results — Method

Collect model run results for a FarmZone

Agtor.component_relationship — Method

Create relational mapping between components and values.

Agtor.extract_values — Method

Extract parameter values from AgParameter

Agtor.generate_agparams — Method

Generate AgParameter definitions.

Arguments

prefix : str

dataset : Dict, of parameters for given component

Returns

Dict matching structure of dataset

Agtor.min_max — Method

min_max(p::AgParameter)
min_max(dataset::Dict)::Dict

Returns min/max of parameter values.

Agtor.run_model — Method

Run scenario for an entire basin.

Agtor.run_model — Method

run_model(zone::FarmZone; ts_func::Function=run_timestep!, pre::Union{Function, Nothing}=nothing, post::Union{Function, Nothing}=nothing)::NamedTuple

Run model for a single zone.

Arguments

zone : Zone AgComponent

tsfunc : Function, defining actions for a time step. Must accept a Manager, zone, int, and Date/DateTime `tsfunc(manager, zone, idx, dt)` All operations must be in-place as changes will not propagate.

pre : Function, defines additional actions for zone that occur at end of time step dt_i: callback_func(zone, dt_i)

post : Function, defines additional actions for zone that occur at end of time step dt_i: callback_func(zone, dt_i)

Returns

NamedTuple : results

Agtor.run_scenarios! — Method

run_scenarios!(samples::DataFrame, basin::Basin, ts_func::Function; 
               pre::Union{Function, Nothing}=nothing, post::Union{Function, Nothing}=nothing)::Dict

Run basin level scenarios for a given sample set.

Returns

Dict with keys: "scenarioid/zoneresults" = zone level results for each zone in basin "scenarioid/fieldresults" = field level results for each zone in basin

Agtor.run_scenarios! — Method

run_scenarios!(samples::DataFrame, zone::FarmZone; ts_func::Function=run_timestep!,
               pre::Union{Function, Nothing}=nothing, post::Union{Function, Nothing}=nothing)::Dict

Run zone level scenarios for a given sample set.

Returns

Dict with keys: - "scenarioid/zoneresults" = zone level results - "scenarioid/fieldresults" = field level results

Agtor.run_timestep! — Method

Run timestep for all zones within a basin.

Agtor.save_results! — Method

Save results for a scenario, basin, and zone combination

Agtor.save_results! — Method

Save arbitrary results to JLD.

Agtor.save_results! — Method

Save results to JLD2.

Example

output_fn = "example.jld2"
results = run_scenarios!(samples, z1, run_timestep!; post=allocation_callback!)
save_results!(output_fn, results)

Agtor.save_results! — Method

Save arbitrary results to JLD.

Agtor.save_results! — Method

Save results for a single zone to JLD.

Agtor.save_state! — Method

Save state to JLD

Agtor.scenario_stats — Method

Get aggregate statistics for an entire scenario run.

Agtor.select_results — Method

Filter DataFrame down to a subset of results based on partial column name match.

Agtor.set_params! — Method

set_params!(comp, sample)
set_params!(comp::Array, sample)::Nothing
set_params!(comp::Dict, sample)::Nothing
set_params!(p::AgParameter, sample::Union{DataFrame, DataFrameRow})::Nothing
set_params!(p::AgParameter, sample::Union{Number, String})::Nothing
set_params!(p::AgParameter, sample::Union{Dict, NamedTuple})::Nothing

Note

update_model!(comp, sample) is an alias for set_params!(comp, sample)

Example

zone_dir = "data_dir/zones/"
zone_specs = load_yaml(zone_dir)
zone_params = generate_agparams("", zone_specs["Zone_1"])

collated_specs = []
collect_agparams!(zone_params, collated_specs; ignore=["crop_spec"])

# Expect only CSV for now...
climate_fn::String = "data/climate/farm_climate_data.csv"
climate::Climate = load_climate(climate_fn)
z1 = create(zone_params, climate)

param_info = DataFrame(collated_specs)

# Generate dataframe of samples
samples = sample(param_info, 1000, sampler)  # where sampler is some function

# Update z1 components with values found in first row
set_params!(z1, samples[1, :])

Agtor.subtotal_costs — Method

Calculate sub-total of irrigation costs

Agtor.total_costs — Function

Calculate total costs for a field.

Maintenance costs can be spread out across a number of fields if desired.

Agtor.update_available_water! — Method

Update water allocations

Agtor.water_used_by_source — Method

water_used_by_source(zone::FarmZone)::DataFrame
water_used_by_source(zone::FarmZone, dt)::DataFrame

Identify source of water used by a zone.