Agtor.jl Documentation

Agtor.BaseManager — Type

An 'economically rational' crop farm manager.

Water is applied for optimal farm profitability based on soil water deficit, crop water requirements, and cost of water application.

Agtor.CategoricalParameter — Type

Categorical parameters.

Min and max values will map to (integer) element position in an CategoricalArray.

Sampling between the min and max values will be mapped to their categorical value in the given array
using `floor()` of the Float value `x`.

Valid values for the CategoricalParameter will therefore be: 
`1 <= x < (n+1)`, where n is number of options.

Agtor.Climate — Type

Serves as an interface to climate data

Agtor.Infrastructure — Type

Represents generic farm infrastructure.

Agtor.Irrigation — Type

On-farm irrigation infrastructure component

Agtor.Pump — Type

Pump(pump_efficiency=0.7, cost_per_kW=0.28, derating=0.75)

An on-farm pump to access water.

Agtor.RigidManager — Type

A rigid farm manager.

Water is applied by a pre-determined quantity rather than optimal profitability. Uses surface water first, then groundwater.

Agtor.ML_water_application_cost — Method

ML_water_application_cost(m::Manager, zone::FarmZone, field::FarmField, req_water_ML_ha::Float64)::NamedTuple

Calculate water application cost/ML by each water source.

Returns

NamedTuple[str, float] : water source name and cost per ML

Agtor.add_prefix! — Method

Modify AgParameter name in place by adding a prefix.

Agtor.apply_irrigation! — Method

Apply irrigation water to field

Agtor.apply_rainfall! — Method

Apply rainfall and ET to influence soil water deficit.

Agtor.apply_rainfall! — Method

Apply rainfall and ET to influence soil water deficit.

Agtor.assign_managers! — Method

Example:

# Load zone specifications
zone_specs = load_spec("test/data/zones")

# Define managers of interest
Manager_A = BaseManager("optimizing")
Manager_B = RigidManager("rigid", 0.05)

# Create tuple-based relation between managers and zones (by name)
manage_zones = ((Manager_A, ("Zone_1", )), (Manager_B, ("Zone_2", "Zone_3")))

# Attach managers to their associated zones
assign_managers!(manage_zones, values(zone_specs))

Raises ArgumentError if duplicate zone names are found.

Agtor.calc_ML_pump_costs — Method

Calculate pumping costs (per ML) for each water source.

Parameters

m : manager component

zone : FarmZone

flowrateLps : float, desired flow rate in Litres per second.

Returns

dict[str, float] : cost of pumping per ML for each water source

Agtor.calc_potential_crop_yield — Method

calc_potential_crop_yield(ssm_mm::Float64, gsr_mm::Float64, crop::AgComponent)

Potential crop yield calculation based on a modified French-Schultz equation. The implemented method is the farmer-friendly version as described by Oliver et al., (2008) (see [1]).

$YP = (SSM + GSR - E) * WUE$

where

$YP$ is yield potential in tonnes per hectare

$SSM$ is Stored Soil Moisture (at start of season) in mm, assumed to be 30% of summer rainfall

$GSR$ is Growing Season Rainfall in mm

$E$ is Crop Evaporation coefficient in mm, the amount of rainfall required before the crop will start to grow, commonly 110mm, but can range from 30-170mm (see [2]),

$WUE$ is Water Use Efficiency coefficient in kg/mm

References

Oliver et al. (2008) (see Equation 1)

Whitbread and Hancock (2008)

Parameters

ssm_mm : float, Stored Soil Moisture (mm) at start of season.

gsr_mm : float, Growing Season Rainfall (mm)

crop : object, Crop component object

Returns

Potential yield in tonnes/Ha

Agtor.calc_required_water — Method

calc_required_water(f::FarmField, dt::Date)

Volume of water to maintain moisture above net irrigation depth (nid).

Calculates volume of water needed to replenish soil water deficit (swd) when SWD falls below nid, considering irrigation efficiency.

Values are given in mm.

link to nid(f::FarmField, dt::Date)

Agtor.collate_field_logs — Method

Collate logged values, aggregating to the zonal level based on identical datetimes.

Agtor.collate_field_logs — Method

Collate logged values, summing on identical datetimes

Agtor.collate_field_logs — Method

Collate logged values, aggregating to the zonal level based on identical datetimes.

Agtor.collect_agparams! — Function

Extract parameter values from AgComponent.

Agtor.collect_agparams — Method

Extract parameter values from Dict specification.

Agtor.collect_results — Method

Collect model run results for a FarmZone

Agtor.ensure_date — Method

Converts strings to datetime.

Agtor.estimate_income_per_ha — Method

Naive estimation of net income.

Agtor.extract_values — Method

Extract parameter values from AgParameter

Agtor.generate_agparams — Method

Generate AgParameter definitions.

Parameters

prefix : str

dataset : Dict, of parameters for given component

Returns

Dict matching structure of dataset

Agtor.get_optimum_irrigated_area — Method

Extract total irrigated area from OptLang optimized results.

Agtor.get_season_range — Method

Gets climate data for season range.

Parameters

p_start : datetime, start of range in Y-m-d format, inclusive.

p_end : datetime, end of range in Y-m-d format, inclusive.

Agtor.get_seasonal_et — Method

Retrieve seasonal evapotranspiration.

Parameters

season_range : List-like, start and end dates, can be string or datetime object

partial_name : str, string to (partially) match column name identifier on

Returns

numeric of seasonal rainfall

Agtor.get_seasonal_rainfall — Method

Retrieve seasonal rainfall by matching column name. Columns names are expected to have 'rainfall' with some identifier.

Parameters

season_range : List-like, start and end dates, can be string or datetime object

partial_name : str, string to (partially) match column name identifier on

Example

Where column names are: 'rainfallfield1', 'rainfallfield2', ...

get_seasonal_rainfall(c, ['1981-01-01', '1982-06-01'], 'field1')

Returns

numeric, representing seasonal rainfall

Agtor.in_season — Method

Check to see if given date is between start and end dates (exclusive of start and end)

Agtor.irrigated_area — Method

The total area marked for irrigation.

Agtor.is_const — Method

Checks if parameter is constant.

Agtor.log_irrigation_by_water_source — Method

Log irrigation volumes from water sources

Agtor.log_irrigation_cost — Method

Log the (total) cost of irrigation.

Agtor.maintenance_cost — Method

Calculate maintenance costs.

Warning: This can be on a per hectare basis or given as a total.

Agtor.matching_dates — Method

Check to see if two dates are identical

Agtor.min_max — Method

Returns min/max values

Agtor.min_max — Method

Returns min/max values

Agtor.nid — Method

nid(f::FarmField, dt::Date)

Calculate net irrigation depth in mm, 0.0 or above.

Equation taken from Agriculture Victoria

See also:

http://www.fao.org/docrep/x5560e/x5560e03.htm

https://www.bae.ncsu.edu/programs/extension/evans/ag452-1.html

http://dpipwe.tas.gov.au/Documents/Soil-waterfactsheet14122011a.pdf

https://www.agric.wa.gov.au/water-management/calculating-readily-available-water?nopaging=1

$NID$ = Effective root depth ($D_{rz}$) $*$ Readily Available Water ($RAW$)

where:

$D_{rz}$ = $Crop_{root depth} * Crop_{e_rz}$, where $Crop_{root depth}$ is the estimated root depth for current stage of crop (initial, late, etc.) and $Crop_{e rz}$ is the effective root zone coefficient for the crop.

$Crop_{e rz}$ is said to be between 1 and 2/3rds of total root depth

$RAW = p * TAW$, $p$ is depletion fraction of crop, $TAW$ is Total Available Water in Soil

As an example, if a crop has a root depth ($RD_{r}$) of 1m, an effective root zone ($RD_{erz}$) coefficient of 0.55, a depletion fraction (p) of 0.4 and the soil has a $TAW$ of 180mm:

$(RD_{r} * RD_{erz}) * (p * TAW)$

Works out to be:

$(1 * 0.55) * (0.4 * 180)$

Returns

* float : net irrigation depth

Agtor.optimize_irrigated_area — Method

Apply Linear Programming to naively optimize irrigated area.

Occurs at start of season.

Parameters

m : Agtor.Manager,

zone : Agtor.FarmZone, representing a farm or a farming zone.

Agtor.optimize_irrigation — Method

optimize_irrigation(m::Manager, zone::FarmZone, dt::Date)::Tuple{OrderedDict, NamedTuple}

Apply Linear Programming to optimize irrigation water use.

Results can be used to represent percentage mix e.g. if the field area is 100 ha, and the optimal area to be irrigated by a water source is

SW: 70 ha
GW: 30 ha

and the required amount is 20mm

SW: 70 / 100 = 0.7 (irrigated area / total area, 70%)
GW: 30 / 100 = 0.3 (30%)

Then the per hectare amount to be applied from each water source is calculated as:

SW = 20mm * 0.7
   = 14mm

GW = 20mm * 0.3
   = 6mm

Parameters

zone : FarmZone

dt : datetime object, current datetime

Returns

Tuple : OrderedDict[str, float] keys based on field and water source names values are hectare area Float : /ML cost of applying water

Agtor.perc_irrigation_sources — Method

Calculate percentage of area to be watered by a specific water source.

Returns

Dict[str, float] : name of water source as key and percent area as value

Agtor.possible_area_by_allocation — Method

Determine the possible irrigation area using water from each water source.

Agtor.possible_irrigation_area — Method

Possible irrigation area in hectares.

Agtor.pumping_costs_per_ML — Method

pumping_costs_per_ML(p::Pump, flow_rate_Lps::Float64, head_pressure::Float64)::Float64

Calculate pumping cost per ML for a given flow rate and head pressure.

$P(Kw) = (H * Q) / ((102 * Ep) * D)$

where:

H is head pressure in metres (m)

Q is Flow in Litres per Second

Ep is Pump efficiency (defaults to 0.7)

D is the derating factor

102 is a constant, as given in Velloti & Kalogernis (2013)

See: * Robinson, D. W. (2002) * Vic. Dept. Agriculture (2006) * Vellotti & Kalogernis (2013)

Parameters

flow_rate_Lps : required flow rate in Litres per second over the irrigation duration

head_pressure : Head pressure of pumping system in metres.

Parameters taken from Pump

pump_efficiency : Efficiency of pump. Defaults to 0.7 (70%)

derating : Accounts for efficiency losses between the energy required at the pump shaft and the total energy required. Defaults to 0.75

Returns

float, costperML

link to Pump(pump_efficiency, cost_per_kW, derating)

Agtor.save_results! — Method

Save results for a scenario, basin, and zone combination

Agtor.save_results! — Method

Save results for a single zone to JLD.

Agtor.save_results! — Method

Save results for a single zone to JLD.

Agtor.seasonal_field_log! — Method

Log seasonal results.

Agtor.set_next_crop! — Method

Sets next crop in rotation and updates sowing/planting dates.

Agtor.set_params! — Method

Usage: zonedir = "datadir/zones/" zonespecs = loadyaml(zonedir) zoneparams = generateagparams("", zonespecs["Zone_1"])

collated_specs = []
collect_agparams!(zone_params, collated_specs; ignore=["crop_spec"])

# Expect only CSV for now...
climate_fn::String = "data/climate/farm_climate_data.csv"
climate::Climate = load_climate(climate_fn)
z1 = create(zone_params, climate)

param_info = DataFrame(collated_specs)

# Generate dataframe of samples
samples = sample(param_info, 1000, sampler)  # where sampler is some function

# Update z1 components with values found in first row
set_params!(z1, samples[1])

Agtor.total_costs — Function

Calculate total costs for a field.

Maintenance costs can be spread out across a number of fields if desired.

Agtor.total_income — Method

total_income(f::FarmField, ssm::Float64, gsr::Float64, irrig::Float64, comps)::Tuple

Calculate net income considering crop yield and costs incurred.

Returns

Tuple:
    * float : net income
    * float : irrigated crop yield, tonnes/Ha
    * float : dryland crop yield, tonnes/Ha

Agtor.update_SWD! — Method

update_SWD!(f::FarmField, rainfall::Float64, ET::Float64)

Calculate soil water deficit.

Water deficit is represented as positive values.

rainfall and ET parameters are expected in mm

Agtor.update_available_water! — Method

Update water allocations

Agtor.update_stages! — Method

Update growth stages with corresponding dates from given sowing date.

Agtor.use_allocation! — Method

Use allocation volume from a particular water source.

Volumes in ML.

Agtor.value_range — Method

Returns max - min, or 0.0 if no min value defined

Agtor.volume_used_by_source — Method

Volume used from a water source in ML

Base.getproperty — Method

Getter for Field

Base.getproperty — Method

Getters for Irrigation

Base.setproperty! — Method

Setter for CategoricalParameter to handle float to categorical element position.

Base.setproperty! — Method

Setter for Field

Agtor.jl Documentation